available places endpoints and place crud had completed, reservation crd operations had completed too
fixes, mock sql files

+++ core-pra/src/main/java/re1kur/core/dto/PlaceDto.java
@@ -2,10 +2,12 @@ package re1kur.core.dto;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
 public record PlaceDto(
         Integer number,
-        Float latitude,
-        Float longitude
+        BigDecimal latitude,
+        BigDecimal longitude
 ) {
 }
